Refactor ESLint rules and configurations: consolidate, update imports, and remove deprecated files
